// src/enums.ts
export enum SizeEnums {
	XXS = '8px',
	XS = '10px',
	S = '12px',
	M = '14px',
	L = '16px',
	XL = '20px',
	XXL = '24px',
	XXXL = '30px',
}

export enum SpaceEnums {
	XXXXS = '2px',
	XXXS = '4px',
	XXS = '6px',
	XS = '8px',
	S = '12px',
	M = '16px',
	L = '20px',
	XL = '24px',
	XXL = '32px',
	XXXL = '36px',
}